Move routing logic to main.jsx and improve authentication flow
